# openapi.yaml
openapi: 3.0.3
info:
  title: New Year 2026 API
  description: |-
    New Year 2026 API Documentation
  contact:
    email: kotaro.otaka@example.com
  version: 0.0.1
servers:
  - url: http://localhost:8080
...
paths:
...
  /.well-known/openid-configuration:
    get:
      tags:
        - OIDC
      summary: OpenID Provider Configuration
      description: OpenID Provider Configuration
      operationId: openIDConfiguration
      security: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenIDConfiguration"
        default:
          description: Internal Server Error
  /authorize:
    get:
      tags:
        - OIDC
      summary: Authorization Endpoint
      description: |
        Authorization Endpoint
        ref: https://openid-foundation-japan.github.io/rfc6749.ja.html#code-authz-req
      operationId: authorize
      security: []
      parameters:
        - name: response_type
          in: query
          required: true
          description: response_type
          schema:
            type: string
            enum:
              - code
        - name: client_id
          in: query
          required: true
          description: client_id
          schema:
            type: string
            format: uuid
            example: 123e4567-e89b-12d3-a456-426614174000
        - name: redirect_uri
          in: query
          required: false
          description: redirect_uri
          schema:
            type: string
            format: uri
            example: http://localhost:8080/callback
        - name: scope
          in: query
          required: false
          description: openid
          schema:
            type: string
            enum:
              - openid
        - name: state
          in: query
          required: false
          description: state
          schema:
            type: string
      responses:
        "302":
          description: Found
          headers:
            Set-Cookie:
              description: Set-Cookie Header
              schema:
                type: string
                example: __authorize__=abcd1234; Path=/; HttpOnly; Secure; SameSite=Strict
            Location:
              schema:
                type: string
                format: uri
                example: http://localhost:8080/callback
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: string
                enum:
                  - invalid_request
                  - unauthorized_client
                  - access_denied
                  - unsupported_response_type
                  - invalid_scope
                  - server_error
                  - temporarily_unavailable
                example: invalid_request
        default:
          description: Internal Server Error
  /callback:
    get:
      tags:
        - OIDC
      summary: Callback Endpoint
      description: Callback Endpoint
      operationId: callback
      security: []
      parameters:
        - name: __authorize__
          in: cookie
          description: auth session
          required: true
          schema:
            type: string
        - name: __session__
          in: cookie
          description: user session
          required: true
          schema:
            type: string
      responses:
        "302":
          description: Found
          headers:
            Set-Cookie:
              description: Set-Cookie Header
              schema:
                type: string
                example: __authorize__=; Path=/; HttpOnly; Secure; SameSite=Strict; Max-Age=-1
            Location:
              schema:
                type: string
                format: uri
                example: https://example.com/callback
        default:
          description: Internal Server Error
  /token:
    post:
      tags:
        - OIDC
      summary: Token Endpoint
      description: |
        Token Endpoint <br>
        ref: https://openid-foundation-japan.github.io/rfc6749.ja.html#token-req <br>
        ref: https://openid-foundation-japan.github.io/openid-connect-core-1_0.ja.html#TokenRequest <br>
      operationId: token
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/TokenRequest"
      responses:
        "200":
          description: OK
          headers:
            Cache-Control:
              schema:
                type: string
                example: no-store
            Pragma:
              schema:
                type: string
                example: no-cache
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenError"
        default:
          description: Internal Server Error
  /certs:
    get:
      tags:
        - OIDC
      summary: JWK Set
      description: JWK Set
      operationId: jwks
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CertsResponse"
        default:
          description: Internal Server Error
  /userinfo:
    get:
      tags:
        - OIDC
      summary: UserInfo Endpoint
      description: UserInfo Endpoint
      operationId: userinfo
      security:
        - Bearer: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfoResponse"
        "401":
          description: Unauthorized
        default:
          description: Internal Server Error
...
  /assertion:
    description: https://developer.mozilla.org/en-US/docs/Web/API/Web_Authentication_API/Attestation_and_Assertion#assertion
    get:
      tags:
        - WebAuthn
      summary: Initialize Assertion
      description: Initialize Assertion
      operationId: initializeAssertion
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {}
          headers:
            Set-Cookie:
              description: Set-Cookie Header
              schema:
                type: string
        default:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - WebAuthn
      summary: Finalize Assertion
      description: Finalize Assertion
      operationId: finalizeAssertion
      parameters:
        - name: __assertion__
          in: cookie
          description: session
          required: true
          schema:
            type: string
            example: session
      requestBody:
        content:
          application/json:
            schema: {}
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
          headers:
            Set-Cookie:
              description: Set-Cookie Header
              schema:
                type: string
        default:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    OpenIDConfiguration:
      type: object
      description: https://openid.net/specs/openid-connect-discovery-1_0.html#ProviderMetadata
      properties:
        issuer:
          type: string
          description: http://localhost:8787
          format: uri
          example: http://localhost:8787
        authorization_endpoint:
          type: string
          description: http://localhost:8787/authorize
          format: uri
          example: http://localhost:8787/authorize
        token_endpoint:
          type: string
          description: http://localhost:8787/token
          format: uri
          example: http://localhost:8787/token
        userinfo_endpoint:
          type: string
          description: http://localhost:8787/userinfo
          format: uri
          example: http://localhost:8787/userinfo
        jwks_uri:
          type: string
          description: http://localhost:8787/certs
          format: uri
          example: http://localhost:8787/certs
        response_types_supported:
          type: array
          items:
            type: string
          example:
            - code
        subject_types_supported:
          type: array
          items:
            type: string
          example:
            - public
        id_token_signing_alg_values_supported:
          type: array
          items:
            type: string
          example:
            - RS256
      required:
        - issuer
        - authorization_endpoint
        - token_endpoint
        - userinfo_endpoint
        - jwks_uri
        - response_types_supported
        - subject_types_supported
        - id_token_signing_alg_values_supported
    TokenRequest:
      type: object
      properties:
        grant_type:
          type: string
          enum:
            - authorization_code
            - refresh_token
          description: grant_type
          example: authorization_code
        code:
          type: string
          description: needs when grant_type is authorization_code
          example: xxxxxxxxxxxxxxx
        redirect_uri:
          type: string
          description: neews when grant_type is authorization_code
          format: uri
          example: http://localhost:3000/callback
        client_id:
          type: string
          description: needs when grant_type is authorization_code
          example: client_id
        client_assertion:
          type: string
          description: needs when grant_type is authorization_code. ref https://datatracker.ietf.org/doc/html/rfc7523#section-2.2
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
        refresh_token:
          type: string
          description: needs when grant_type is refresh_token
          example: xxxxxxxx
        scope:
          type: string
          description: optional
          example: openid profile email
      required:
        - grant_type
        # - code # when grant_type is authorization_code. not required if grant_type is refresh_token
    TokenResponse:
      type: object
      description: https://openid-foundation-japan.github.io/openid-connect-core-1_0.ja.html#TokenResponse
      properties:
        access_token:
          type: string
          description: access_token
          example: access_token
        token_type:
          type: string
          description: token_type
          example: token_type
        refresh_token:
          type: string
          description: refresh_token
          example: refresh_token
        expires_in:
          type: integer
          description: expires_in
          example: 3600
        id_token:
          type: string
          description: id_token
          example: id_token
      required:
        - access_token
        - token_type
        - refresh_token
        - expires_in
        - id_token
    TokenError:
      type: object
      properties:
        error:
          type: string
          enum:
            - invalid_request
            - invalid_client
            - invalid_grant
            - unauthorized_client
            - unsupported_grant_type
            - invalid_scope
          example: invalid_request
        error_description:
          type: string
        error_uri:
          type: string
      required:
        - error
        - error_description
        - error_uri
    CertsResponse:
      type: object
      properties:
        keys:
          type: array
          items:
            $ref: "#/components/schemas/JWKSet"
      required:
        - keys
    JWKSet:
      type: object
      properties:
        kid:
          type: string
          example: xxxxxxxxxx
        kty:
          type: string
          example: RSA
        use:
          type: string
          example: sig
        alg:
          type: string
          example: RS256
        n:
          type: string
          example: xxxxxxxxxx
        e:
          type: string
          example: xxxxxxxxxx
      required:
        - kid
        - kty
        - use
        - alg
        - n
        - e
    UserInfoResponse:
      type: object
      properties:
        sub:
          type: string
          example: xxxxxxxxxx
        email:
          type: string
          example: xxxxxxxxxx
      required:
        - sub
        - email
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT